requirements.txt
gym-retro
opencv-python
pynput
pandas